Deploy kubernetes-nmstate
This PR change the way we deploy kubernetes-nmstate.

Before this PR we use the cluster-network-addnos-operator.
We plan to add nmstate as part of the openshift network operator in the future.

To have an easy upgrade after the openshift network operator is better for now
to deploy nmstate without an operator.

Signed-off-by: Sebastian Sch <sebassch@gmail.com>