manifests: add a way to generate single manifests
And expose it in the testsuite, to be abile to easily parametrize
the generation and check it.
Unfortunately we will need a minimal difference between the single
manifest flow, currently used in tests, and the all manifests flow,
currently used to deliver the final manifests.

We should eliminate this difference to make the testing step
even more reliable

Signed-off-by: Francesco Romani <fromani@redhat.com>