Fix incorrect URL for openshift-client archive (#392)
In "Retrieving OpenShift Installer GA" [1], we see the following instructions:

    . Get the `oc` binary:
    +
    [source,bash]
    ----
    [kni@provisioner ~]$ curl -s https://mirror.openshift.com/pub/openshift-v4/clients/ocp/$VERSION/openshift-client-linux-$VERSION.tar.gz | tar zxvf - oc
    ----

Unfortunately, at this point, `$VERSION` is set to `latest-4.4`.  This
results in the following URL:

    https://mirror.openshift.com/pub/openshift-v4/clients/ocp/latest-4.4/openshift-client-linux-latest-4.4.tar.gz

That will result in a 404 error, because the files in that directory
have names like `openshift-install-linux-4.4.4.tar.gz` or
`openshift-install-linux.tar.gz`. The easiest solution is to
use the unversioned filename.

[1]: https://openshift-kni.github.io/baremetal-deploy/4.4/Deployment.html#retrieving-openshift-installer-ga