Clean up nmcli scripts
The existing `nmcli` script jumbled together commands for the
baremetal and provisioning network, making it difficult to understand
exactly what was going on. There were also some commands that were
effectively "hidden" after a semicolon which weren't immediately
obvious when inspecting the script.

This changes splits up the nmcli scripts into two commands (one for the baremetal
network, one for the provisioning network), and replaces the direct
manipulation of `dhclient` with `nmcli con down` and `nmcli con up`
commands.