Per https://bugzilla.redhat.com/show_bug.cgi?id=1923965, I modified BMC docs.
The BZ calls for an address table with the vendors. I thought it might make more sense to segregate the vendors into different modules to reduce the scope for the reader, and also include the table. I also re-ordered it so that the most favored approach--RedFish Virtual Media--is first and IPMI is last since it isn't secure.

Fixes: BZ-1923965

Signed-off-by: John Wilkins <jowilkin@redhat.com>