Added support for redfish for KVM hosts
Currently redfish is supported for HP and Dell systems, but none of
these vendors provide settings compatible with sushy-tools based virtual
BMCs. For instance, all the KVMs in the same hypervisor share the BMC
address, and are differentiated by the system ID, which is part of the
resource path.

This change extends the install-config and install-config-virtualmedia
templates to allow for sushy-tools based redfish vBMCs.