Allowing status code 301 when verifying the internet connectivity
Currently, when checking the connectivity to the Internet, an HTTP query
is set against https://www.redhat.com.

This address will respond with a 301 status code redirecting the client
to https://www.redhat.com/en.

The URI ansible module is able to follow redirects so, if everything is
OK, the query shour result in either a 200 status code if the server is
reachable, or a -1 if it's not. In this later scenario, a disconnected
environment is assumed.

The problem is in some labs the web server gets into an infinite
redirect loop, so the resulting status code would be 301.

This change allows for this status code to be accepted as evidence that
the Red Hat web site was reached and, therefore, the network has
Internet access.